requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dj-database-url==0.5.0
Django==2.2.9
django-crispy-forms==1.8.1
django-heroku==0.3.1
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-speech==2.4.0
google-cloud-storage==1.38.0
google-cloud-vision==2.3.1
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpc-google-cloud-vision-v1==0.14.0
grpcio==1.38.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
Jinja2==2.10.3
libcst==0.3.19
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==3.0.0
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.17.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
whitenoise==5.2.0
